+++ expo-jest/resources/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  preset: "jest-expo",
+  transformIgnorePatterns: [
+    "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)",
+  ],
+  moduleFileExtensions: ["ts", "tsx", "js", "ts"],
+  collectCoverage: true,
+};
